added interactive map, restructured folder tree
