+++ client/src/pages/ContactUsPage/ContactUsPage.css
@@ -2,7 +2,7 @@
 	margin: 0 20px;
 }
 
-@media screen and (min-width: 700px) {
+@media screen and (min-width: 768px) {
 	.contact-us-page {
 	}
 }
