+++ client/src/pages/ContactUsPage/components/VisitUsMap/VisitUsMap.css
@@ -1,0 +1,11 @@
+.visit-us-map-container {
+	width: 100%;
+	height: 400px;
+	border: 1px solid var(--black);
+}
+
+@media screen and (min-width: 768px) {
+	.visit-us-map-container {
+		height: auto;
+	}
+}
