--- client/src/pages/ContactUsPage/components/VisitUsSection/VisitUsMap/VisitUsMap.css
@@ -1,30 +1,0 @@
-/* i am not happy with the way this image and it's container is working */
-/* i want it zoomed in at 1:1 pixel ratio and centered to the center point of the image */
-
-.visit-us-map-container {
-	position: relative;
-	width: 100%;
-	height: 300px;
-	/* different aspect ratio for Mobile */
-	/* padding-bottom: 90%; */
-}
-
-.visit-us-map {
-	position: absolute;
-	width: 100%;
-	height: 100%;
-	object-fit: cover;
-	object-position: center;
-	border: 1px solid black;
-	background-color: hsl(0, 0%, 70%);
-}
-
-@media screen and (min-width: 700px) {
-	.visit-us-map-container {
-		/* different aspect ratio for Desktop */
-		/* padding-bottom: 55%; */
-	}
-
-	.visit-up-map {
-	}
-}
