--- client/src/pages/ContactUsPage/components/VisitUsSection/VisitUsMap/VisitUsMap.jsx
@@ -1,17 +1,0 @@
-import "./VisitUsMap.css";
-import mapImage from "./map.png";
-
-const VisitUsMap = () => {
-	return (
-		<div className="visit-us-map-container">
-			{/* this should eventually be a Map API component */}
-			<img
-				src={mapImage}
-				alt="aerial photograph of the Lentegeur Hospital area"
-				className="visit-us-map"
-			/>
-		</div>
-	);
-};
-
-export default VisitUsMap;
