+++ client/src/pages/ContactUsPage/components/VisitUsSection/VisitUsSection.css
@@ -36,7 +36,7 @@
 	font-size: inherit;
 }
 
-@media screen and (min-width: 700px) {
+@media screen and (min-width: 768px) {
 	.visit-us-container {
 		width: 90%; /* the map is flush with the right edge but i've changed it deliberately */
 		margin: 0 auto;
